DPL: add ability to adjust topology from services
Notice that for the moment services are loaded in any case.
