+++ Framework/Core/src/ArrowSupport.cxx
@@ -9,9 +9,13 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 #include "ArrowSupport.h"
+
+#include "Framework/AODReaderHelpers.h"
 #include "Framework/ArrowContext.h"
 #include "Framework/DataProcessor.h"
 #include "Framework/ServiceRegistry.h"
+#include "Framework/ConfigContext.h"
+#include "Framework/CommonDataProcessors.h"
 #include "Framework/DeviceSpec.h"
 #include "Framework/EndOfStreamContext.h"
 #include "Framework/Tracing.h"
@@ -19,6 +23,8 @@
 #include "Framework/DeviceMetricsHelper.h"
 #include "Framework/DeviceInfo.h"
 #include "Framework/DevicesManager.h"
+#include "WorkflowHelpers.h"
+#include "Framework/WorkflowSpecNode.h"
 
 #include "CommonMessageBackendsHelpers.h"
 #include <Monitoring/Monitoring.h>
@@ -382,10 +388,132 @@ o2::framework::ServiceSpec ArrowSupport::arrowBackendSpec()
                          LOGP(INFO, "Rate limiting set up at {}MB distributed over {} readers", config->maxMemory, readers);
                          registry.registerService(ServiceRegistryHelpers::handleForService<RateLimitConfig>(config));
                          once = true;
-                       }
-                     },
-                     nullptr,
-                     ServiceKind::Global};
+                       } },
+                     .adjustTopology = [](WorkflowSpecNode& node, ConfigContext const& ctx) {
+      auto workflow = node.specs;
+      auto spawner = std::find_if(workflow.begin(), workflow.end(), [](DataProcessorSpec& spec) { return spec.name == "internal-dpl-aod-spawner"; });
+      auto builder = std::find_if(workflow.begin(), workflow.end(), [](DataProcessorSpec& spec) { return spec.name == "internal-dpl-aod-index-builder"; });
+      auto reader = std::find_if(workflow.begin(), workflow.end(), [](DataProcessorSpec& spec) { return spec.name == "internal-dpl-aod-reader"; });
+      auto writer = std::find_if(workflow.begin(), workflow.end(), [](DataProcessorSpec& spec) { return spec.name == "internal-dpl-aod-writer"; });
+
+      if (spawner != workflow.end()) {
+        std::vector<InputSpec> requestedDYNs;
+        // collect currently requested DYNs
+        for (auto& d : workflow) {
+          if (d.name == spawner->name) {
+            continue;
+          }
+          for (auto& i : d.inputs) {
+            if (DataSpecUtils::partialMatch(i, header::DataOrigin{"DYN"})) {
+              requestedDYNs.emplace_back(i);
+            }
+          }
+          std::sort(requestedDYNs.begin(), requestedDYNs.end(), [](InputSpec const& a, InputSpec const& b) { return a.binding < b.binding; });
+          auto end = std::unique(requestedDYNs.begin(), requestedDYNs.end(), [](InputSpec const& a, InputSpec const& b) { return a.binding == b.binding; });
+          requestedDYNs.erase(end, requestedDYNs.end());
+        }
+
+        // remove unmatched outputs
+        auto o_end = std::remove_if(spawner->outputs.begin(), spawner->outputs.end(), [&](OutputSpec const& o) {
+          return std::none_of(requestedDYNs.begin(), requestedDYNs.end(), [&](InputSpec const& i) { return DataSpecUtils::match(i, o); });
+        });
+        spawner->outputs.erase(o_end, spawner->outputs.end());
+
+        // remove unmatched inputs
+        auto i_end = std::remove_if(spawner->inputs.begin(), spawner->inputs.end(), [&](InputSpec const& i) {
+          auto&& [origin, description] = DataSpecUtils::asConcreteDataTypeMatcher(i);
+          return std::none_of(spawner->outputs.begin(), spawner->outputs.end(), [description = description](OutputSpec const& o) {
+            return DataSpecUtils::partialMatch(o, description);
+          });
+        });
+        spawner->inputs.erase(i_end, spawner->inputs.end());
+
+        // replace AlgorithmSpec
+        // FIXME: it should be made more generic, so it does not need replacement...
+        spawner->algorithm = readers::AODReaderHelpers::aodSpawnerCallback(requestedDYNs);
+      }
+
+      if (builder != workflow.end()) {
+        // collect currently requested IDXs
+        std::vector<InputSpec> requestedIDXs;
+        std::vector<InputSpec> dummy_i;
+        std::vector<OutputSpec> dummy_o;
+        for (auto& d : workflow) {
+          if (d.name == builder->name) {
+            continue;
+          }
+          for (auto& i : d.inputs) {
+            if (DataSpecUtils::partialMatch(i, header::DataOrigin{"IDX"})) {
+              requestedIDXs.emplace_back(i);
+            }
+          }
+          std::sort(requestedIDXs.begin(), requestedIDXs.end(), [](InputSpec const& a, InputSpec const& b) { return a.binding < b.binding; });
+          auto end = std::unique(requestedIDXs.begin(), requestedIDXs.end(), [](InputSpec const& a, InputSpec const& b) { return a.binding == b.binding; });
+          requestedIDXs.erase(end, requestedIDXs.end());
+        }
+
+        // recreate inputs and outputs
+        builder->inputs = dummy_i;
+        builder->outputs = dummy_o;
+        auto copy = requestedIDXs;
+        WorkflowHelpers::addMissingOutputsToBuilder(std::move(copy), dummy_i, *builder);
+
+        //replace AlgorithmSpec
+        // FIXME: it should be made more generic, so it does not need replacement...
+        builder->algorithm = readers::AODReaderHelpers::indexBuilderCallback(requestedIDXs);
+      }
+
+      if (reader != workflow.end() && (spawner != workflow.end() || builder != workflow.end())) {
+        // If reader and/or builder were adjusted, remove unneeded outputs
+        std::vector<InputSpec> requestedAODs;
+        // collect currently requested AODs
+        for (auto& d : workflow) {
+          for (auto& i : d.inputs) {
+            if (DataSpecUtils::partialMatch(i, header::DataOrigin{"AOD"})) {
+              requestedAODs.emplace_back(i);
+            }
+          }
+          std::sort(requestedAODs.begin(), requestedAODs.end(), [](InputSpec const& a, InputSpec const& b) { return a.binding < b.binding; });
+          auto end = std::unique(requestedAODs.begin(), requestedAODs.end(), [](InputSpec const& a, InputSpec const& b) { return a.binding == b.binding; });
+          requestedAODs.erase(end, requestedAODs.end());
+        }
+
+        // remove unmatched outputs
+        auto o_end = std::remove_if(reader->outputs.begin(), reader->outputs.end(), [&](OutputSpec const& o) {
+          return !DataSpecUtils::partialMatch(o, o2::header::DataDescription{"TFNumber"}) && std::none_of(requestedAODs.begin(), requestedAODs.end(), [&](InputSpec const& i) { return DataSpecUtils::match(i, o); });
+        });
+        reader->outputs.erase(o_end, reader->outputs.end());
+      }
+
+      if (writer != workflow.end()) {
+        workflow.erase(writer);
+        // replace writer as some outputs may have become dangling and some are now consumed
+        auto [outputsInputs, outputTypes] = WorkflowHelpers::analyzeOutputs(workflow);
+
+        // create DataOutputDescriptor
+        std::shared_ptr<DataOutputDirector> dod = WorkflowHelpers::getDataOutputDirector(ctx.options(), outputsInputs, outputTypes);
+
+        // select outputs of type AOD which need to be saved
+        // ATTENTION: if there are dangling outputs the getGlobalAODSink
+        // has to be created in any case!
+        std::vector<InputSpec> outputsInputsAOD;
+        for (auto ii = 0u; ii < outputsInputs.size(); ii++) {
+          if ((outputTypes[ii] & WorkflowHelpers::ANALYSIS) == WorkflowHelpers::ANALYSIS) {
+            auto ds = dod->getDataOutputDescriptors(outputsInputs[ii]);
+            if (!ds.empty() || (outputTypes[ii] & WorkflowHelpers::DANGLING) == WorkflowHelpers::DANGLING) {
+              outputsInputsAOD.emplace_back(outputsInputs[ii]);
+            }
+          }
+        }
+
+        // file sink for any AOD output
+        if (!outputsInputsAOD.empty()) {
+          // add TFNumber as input to the writer
+          outputsInputsAOD.emplace_back(InputSpec{"tfn", "TFN", "TFNumber"});
+          workflow.push_back(CommonDataProcessors::getGlobalAODSink(dod, outputsInputsAOD));
+        }
+      } },
+                     .kind = ServiceKind::Global};
 }
 
 } // namespace o2::framework
