+++ Framework/Core/src/WorkflowHelpers.h
@@ -151,6 +151,12 @@ enum struct WorkflowParsingState : int {
 
 /// A set of internal helper classes to manipulate a Workflow
 struct WorkflowHelpers {
+  enum OutputType : char {
+    UNKNOWN = 0,
+    DANGLING = 1,
+    ANALYSIS = 2,
+  };
+
   /// Topological sort for a graph of @a nodeCount nodes.
   ///
   /// @a edgeIn pointer to the index of the input node for the first edge
@@ -177,6 +183,19 @@ struct WorkflowHelpers {
   // @a workflow the workflow to decorate
   // @a ctx the context for the configuration phase
   static void injectServiceDevices(WorkflowSpec& workflow, ConfigContext const& ctx);
+
+  /// Helper functions to add AOD related internal devices.
+  /// FIXME: moved here until we have proper plugin based amendment
+  ///        of device injection
+  static void addMissingOutputsToReader(std::vector<OutputSpec> const& providedOutputs,
+                                        std::vector<InputSpec> requestedInputs,
+                                        DataProcessorSpec& publisher);
+  static void addMissingOutputsToSpawner(std::vector<InputSpec>&& requestedDYNs,
+                                         std::vector<InputSpec>& requestedAODs,
+                                         DataProcessorSpec& publisher);
+  static void addMissingOutputsToBuilder(std::vector<InputSpec>&& requestedIDXs,
+                                         std::vector<InputSpec>& requestedAODs,
+                                         DataProcessorSpec& publisher);
 
   // Re-adjust service devices if the inputs of other devices were modified
   // @a workflow to analyze
